[Input-Format] Added few changes to the input text common class.
